Added more documentation and moved some files around to make better sense.
