+++ src/buff_manager/models/buffManagerConfig.model.ts
@@ -1,4 +1,4 @@
-import { BasicModel } from "../../shared/models/basicModel.js";
+import { BasicModel } from "../../utils/models/index.js";
 import { Buff } from "./buff.model.js";
 import { MessageSettings } from "./messageSettings.model.js";
 import { Week } from "./week.model.js";
