--- src/config/app_configs/index.ts
@@ -1,4 +1,0 @@
-export * from "./appConfigs.js";
-export * from "./models/commandRegistrationConfiguration.model.js";
-export * from "./models/databaseConfiguration.model.js";
-export * from "./models/projectConfiguration.model.js";
