+++ src/event_manager/eventManager.module.ts
@@ -2,7 +2,7 @@ import { Client, CommandInteraction, Message } from "discord.js";
 import { singleton } from "tsyringe";
 
 import { addCommandKeys, authorize } from "../permission_manager/index.js";
-import { ModuleBase } from "../shared/models/moduleBase.js";
+import { ModuleBase } from "../utils/models/index.js";
 import { EventManagerService } from "./eventManager.service.js";
 
 @singleton()
