+++ src/event_manager/models/eventManagerConfig.model.ts
@@ -1,4 +1,4 @@
-import { BasicModel } from "../../shared/models/basicModel.js";
+import { BasicModel } from "../../utils/models/index.js";
 import { EventObj } from "./eventObj.model.js";
 import { ReminderTrigger } from "./reminderTrigger.model.js";
 import { Tags } from "./tags.model.js";
