+++ src/gardening_manager/gardeningManager.service.ts
@@ -3,9 +3,9 @@ import { DateTime } from "luxon";
 import { pino } from "pino";
 import { singleton } from "tsyringe";
 
-import { createLogger } from "../shared/logger/logger.decorator.js";
-import { Client } from "../shared/models/client.js";
-import { InvalidArgumentError } from "../shared/models/errors.js";
+import { createLogger } from "../utils/logger/logger.decorator.js";
+import { Client } from "../utils/models/client.js";
+import { InvalidArgumentError } from "../utils/models/errors.js";
 import { GardeningManagerRepository } from "./gardeningManager.repository.js";
 import { GardeningConfig, Plot, Reason, Reservation, Slot } from "./models/index.js";
