+++ src/index.ts
@@ -5,7 +5,7 @@ import { program } from "commander";
 import * as os from "os";
 
 import { main } from "./app.js";
-import { commandRegistration } from "./config/commandRegistration.js";
+import { commandRegistration } from "./commandRegistration.js";
 
 console.log(`Welcome ${os.userInfo().username}.`);
