+++ src/permission_manager/permissionManager.module.ts
@@ -2,7 +2,7 @@ import { SlashCommandStringOption } from "@discordjs/builders";
 import { CommandInteraction, Role } from "discord.js";
 import { singleton } from "tsyringe";
 
-import { ModuleBase } from "../shared/models/moduleBase.js";
+import { ModuleBase } from "../utils/models/index.js";
 import { addCommandKeys } from "./addCommandKeys.decorator.js";
 import { authorize } from "./authorize.decorator.js";
 import { PermissionMode } from "./models/index.js";
