+++ src/permission_manager/permissionManager.service.ts
@@ -2,8 +2,8 @@ import { CommandInteraction, Interaction, MessageEmbed, Role } from "discord.js"
 import { pino } from "pino";
 import { singleton } from "tsyringe";
 
-import { createLogger } from "../shared/logger/logger.decorator.js";
-import { deepMerge } from "../shared/utils.js";
+import { createLogger } from "../utils/logger/logger.decorator.js";
+import { deepMerge } from "../utils/utils.js";
 import { PermissionKeys } from "./addCommandKeys.decorator.js";
 import { Permission, PermissionKeysType, PermissionManagerConfig, PermissionMode } from "./models/index.js";
 import { PermissionManagerRepository } from "./permissionManager.repository.js";
