+++ src/role_manager/roleManager.model.ts
@@ -1,4 +1,4 @@
-import { BasicModel } from "../shared/models/basicModel.js";
+import { BasicModel } from "../utils/models/index.js";
 
 export class RoleManagerConfig extends BasicModel {
     public guildId: string;
