+++ src/role_manager/roleManager.service.ts
@@ -2,10 +2,10 @@ import { CommandInteraction, Guild, GuildMember, Message, MessageReaction, React
 import { pino } from "pino";
 import { singleton } from "tsyringe";
 
-import { createLogger } from "../shared/logger/logger.decorator.js";
-import { Client } from "../shared/models/client.js";
-import { Task } from "../shared/models/task.js";
-import { fetchMessages } from "../shared/utils.js";
+import { createLogger } from "../utils/logger/logger.decorator.js";
+import { Client } from "../utils/models/client.js";
+import { Task } from "../utils/models/task.js";
+import { fetchMessages } from "../utils/utils.js";
 import { RoleManagerConfig } from "./roleManager.model.js";
 import { RoleManagerRepository } from "./roleManager.repository.js";
