+++ src/utils/configuration.class.ts
