+++ src/utils/models/basicModel.ts
