+++ src/utils/models/commandRegistrationConfiguration.model.ts
