+++ test/utils/mockDatabase.ts
@@ -1,8 +1,8 @@
 import { Db } from "mongodb";
 import { singleton } from "tsyringe";
 
-import { DatabaseConfiguration } from "../../src/config/databaseConfiguration.js";
-import { deepMerge } from "../../src/shared/utils.js";
+import { DatabaseConfiguration } from "../../src/utils/config/databaseConfiguration.js";
+import { deepMerge } from "../../src/utils/utils.js";
 
 @singleton()
 export class MockDatabase extends DatabaseConfiguration {
